requirements.txt
asgiref
cmake
crispy-bootstrap5
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
pillow
setuptools
sqlparse
tzdata
wheel